% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{bureaucrats}
\alias{bureaucrats}
\title{Village Bureaucrats (CGVO) Dataset}
\format{
A data frame with 2,809 rows and 18 variables:
\describe{
\item{village}{Village ID}
\item{year}{Year}
\item{cgvo}{CGVO has been assigned}
\item{l_subsidy_rate}{Log subsidized population (per 1000)}
\item{l_poor_housing_rate}{Log poor housing households (per 100)}
\item{l_poor_reg_rate}{Log registered poor households (per 100)}
\item{l_disability_rate}{Log population with disabilities (per 1000)}
\item{district}{District ID}
}
}
\source{
He, Guojun, and Shaoda Wang. 2017. "Do College Graduates Serving
as Village Officials Help Rural China?" \emph{American Economic Journal:
Applied Economics} 9(4): 186-215. \doi{10.1257/app.20160079}
}
\description{
Village-year panel studying the effects of assigning college graduate
village officials (CGVOs) to rural Chinese villages.
}
\examples{
\dontrun{
bur <- grossman::load("bureaucrats")
}

}
\keyword{datasets}
