% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{cigs}
\alias{cigs}
\title{Cigarettes Dataset}
\format{
A data frame with 1,209 rows and 7 variables:
\describe{
\item{state}{State}
\item{year}{Year}
\item{cigsales}{Cigarette sales per capita}
\item{gdp}{Log GDP per capita}
\item{price}{Average retail price of cigarettes}
\item{age1524}{Proportion of population aged 15-24}
\item{beer}{Per capita beer consumption}
}
}
\source{
Abadie, Alberto, Alexis Diamond, and Jens Hainmueller. 2010.
"Synthetic Control Methods for Comparative Case Studies: Estimating the
Effect of California's Tobacco Control Program." \emph{Journal of the
American Statistical Association} 105(490): 493-505.
\doi{10.1198/jasa.2009.ap08746}
}
\description{
State-year panel on cigarette sales, prices, and demographics.
}
\examples{
\dontrun{
cigs <- grossman::load("cigs")
}

}
\keyword{datasets}
