% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{cookstove}
\alias{cookstove}
\title{Cookstove Adoption Dataset}
\format{
A data frame with 7,949 rows and 7 variables:
\describe{
\item{SMS_date}{SMS survey date}
\item{respondent_id}{Respondent ID}
\item{amount_weekly}{Weekly charcoal spending}
\item{jikokoa}{Bought cookstove}
\item{wtp}{Willingness to pay (USD)}
\item{amount_weekly_pre}{Pre-treatment weekly charcoal spending}
\item{price}{Offered price (USD)}
}
}
\source{
Berkouwer, Susanna B. and Joshua T. Dean. 2022. "Credit, Attention,
and Externalities in the Adoption of Energy Efficient Technologies by
Low-Income Households." \emph{American Economic Review} 112(10): 3291-3330.
\doi{10.1257/aer.20210766}
}
\description{
SMS survey panel on cookstove adoption in Kenya. Used to study
the role of credit and attention in adoption of energy efficient
technologies.
}
\examples{
\dontrun{
cookstove <- grossman::load("cookstove")
}

}
\keyword{datasets}
