% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{dippel}
\alias{dippel}
\title{Dippel Reservations IV Dataset}
\format{
A data frame with 182 rows and 19 variables:
\describe{
\item{tribeid}{Tribe identifier}
\item{per.capita.income}{Per capita income (USD)}
\item{forced}{Forced coexistence indicator}
\item{gold}{Gold resource endowment}
\item{silver}{Silver resource endowment}
\item{historical.centralization}{Historical political centralization}
\item{county.per.capita.income}{County per capita income (USD)}
\item{county.unemployment}{County unemployment rate}
\item{state}{US state}
}
}
\source{
Dippel, Christian. 2014. "Forced Coexistence and Economic
Development: Evidence from Native American Reservations."
\emph{Econometrica} 82(6): 2131-2165. \doi{10.3982/ECTA11423}
}
\description{
Cross-section of Native American reservations studying the effect of
forced coexistence between indigenous groups on economic development,
using historical precious metal endowments as instruments.
}
\examples{
\dontrun{
dippel <- grossman::load("dippel")
}

}
\keyword{datasets}
