% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{eskom}
\alias{eskom}
\title{Eskom Electrification Dataset}
\format{
A data frame with 1816 rows and 18 variables:
\describe{
\item{eskom}{Treatment indicator: 1 if electrified 1996-2001, 0 otherwise}
\item{gradient}{Average community land gradient (instrument)}
\item{km_to_grid}{Distance from community centroid to nearest electricity grid (km)}
\item{km_to_road}{Distance to nearest road (km)}
\item{km_to_town}{Distance to nearest town (km)}
\item{hh_dens}{Household density in 1996}
\item{hh_pov}{Fraction of households below ZAR 6,000/year in 1996}
\item{pr_female_hh_head}{Fraction of households headed by a female}
\item{sex_ratio}{Ratio of African adult women to men in 1996}
\item{pr_indian_or_white}{Fraction of households headed by Indian or white person}
\item{pr_hs_men}{Fraction of African men with matric (high school)}
\item{pr_hs_women}{Fraction of African women with matric (high school)}
\item{d_water_access}{Change in fraction of households with water source nearby}
\item{d_toilet}{Change in fraction of households with flush toilet}
\item{d_pr_emp_f}{Change in female employment rate (2001 - 1996)}
\item{d_pr_emp_m}{Change in male employment rate (2001 - 1996)}
\item{district}{District code}
\item{subdistrict}{Subdistrict (main place) code}
}
}
\source{
Dinkelman, Taryn. 2011. "The Effects of Rural Electrification on
Employment: New Evidence from South Africa." \emph{American Economic Review}
101(7): 3078-3108. \doi{10.1257/aer.101.7.3078}
}
\description{
Community-level data from South Africa used to study the effects of rural
electrification on employment. The Eskom electrification project (1996-2001)
provides quasi-experimental variation in electricity access.
}
\examples{
\dontrun{
# Load the dataset
eskom <- grossman::load("eskom")

# Replicate Table 3, Column 4 (first stage)
first_stage <- lm(eskom ~ gradient + km_to_grid + hh_dens + hh_pov +
                  pr_female_hh_head + sex_ratio + pr_indian_or_white +
                  km_to_road + km_to_town + pr_hs_men + pr_hs_women +
                  d_water_access + d_toilet + as.factor(district),
                  data = eskom)
summary(first_stage)
}

}
\keyword{datasets}
