% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{grossman_cache_dir}
\alias{grossman_cache_dir}
\title{Get the local cache directory for grossman data}
\usage{
grossman_cache_dir()
}
\value{
Path to cache directory
}
\description{
Get the local cache directory for grossman data
}
\keyword{internal}
