% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{hurricane}
\alias{hurricane}
\title{Hurricane Transfers Dataset}
\format{
A data frame with 49,698 rows and 7 variables:
\describe{
\item{county}{5-digit county FIPS code}
\item{year}{Year}
\item{hyear}{Hurricane year}
\item{govtrans}{Log government transfers per capita}
\item{bustrans}{Log business transfers per capita}
\item{emprate}{CBP employment / population aged 15+}
\item{avgwage}{Log average wage/salary}
}
}
\source{
Deryugina, Tatyana. 2017. "The Fiscal Cost of Hurricanes: Disaster
Aid versus Social Insurance." \emph{American Economic Journal: Economic
Policy} 9(3): 168-198. \doi{10.1257/pol.20140296}
}
\description{
County-year panel on government transfers and economic outcomes around
hurricanes, used to study the fiscal costs of natural disasters.
}
\examples{
\dontrun{
hurr <- grossman::load("hurricane")
}

}
\keyword{datasets}
