% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{kenyagrid}
\alias{kenyagrid}
\title{Kenya Grid Electrification RCT Dataset}
\format{
A data frame with 4,368 rows and 52 variables:
\describe{
\item{hhid}{Household ID}
\item{connected}{Grid connected}
\item{transearly}{Transformer funded early (treatment)}
\item{treated_all}{Any treatment}
\item{emi}{Economic index}
\item{nemi}{Non-economic index}
\item{energy_spending}{Monthly total energy spending (USD)}
}
}
\source{
Lee, Kenneth, Edward Miguel, and Catherine Wolfram. 2020.
"Experimental Evidence on the Economics of Rural Electrification."
\emph{Journal of Political Economy} 128(4): 1523-1565.
\doi{10.1086/705417}
}
\description{
Household panel from a rural electrification randomized controlled trial
in Western Kenya, studying the economic and social impacts of grid access.
}
\examples{
\dontrun{
kg <- grossman::load("kenyagrid")
}

}
\keyword{datasets}
