% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{kindy}
\alias{kindy}
\title{Kindergarten and Maternal Labor Supply Dataset}
\format{
A data frame with 17,817 rows and 23 variables:
\describe{
\item{youngest}{Age of youngest child}
\item{public}{Child enrolled in public school}
\item{quarter}{Quarter of birth}
\item{worked}{Worked last week}
\item{hours}{Hours worked last week}
\item{work79}{Employment in 1979}
\item{hours79}{Hours worked in 1979}
\item{salary}{Wage and salary income in 1979}
}
}
\source{
Gelbach, Jonah B. 2002. "Public Schooling for Young Children and
Maternal Labor Supply." \emph{American Economic Review} 92(1): 307-322.
\doi{10.1257/000282802760015748}
}
\description{
Data on kindergarten enrollment and maternal labor supply outcomes,
using quarter of birth as an instrument for public school enrollment.
}
\examples{
\dontrun{
kindy <- grossman::load("kindy")
}

}
\keyword{datasets}
