% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{pisa}
\alias{pisa}
\title{PISA Incentive Experiment}
\format{
A data frame with 1,103 rows and 21 variables:
\describe{
\item{score}{Test score}
\item{incent}{Treatment: received incentive (1=yes, 0=no)}
\item{us}{United States indicator (1=US, 0=China)}
\item{female}{Female indicator}
\item{group}{Treatment group assignment}
}
}
\source{
Gneezy, Uri, John A. List, Jeffrey A. Livingston, Xiangdong Qin,
Sally Sadoff, and Yang Xu. 2019. "Measuring Success in Education: The Role
of Effort on the Test Itself." \emph{American Economic Review: Insights}
1(3): 291-308. \doi{10.1257/aeri.20180633}
}
\description{
Experimental data testing whether financial incentives improve student
test performance, comparing US and Chinese schools.
}
\examples{
\dontrun{
pisa <- grossman::load("pisa")
}

}
\keyword{datasets}
