% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{queens}
\alias{queens}
\title{Queens and Wars Dataset}
\format{
A data frame with 3,586 rows and 17 variables:
\describe{
\item{kingdom}{Polity name}
\item{year}{Year}
\item{inwar}{In war}
\item{queen}{Queen}
\item{fbmale}{First-born male (of previous monarchs)}
\item{century}{Century}
\item{decade}{Decade}
}
}
\source{
Dube, Oeindrila, and S.P. Harish. 2020. "Queens."
\emph{Journal of Political Economy} 128(7): 2579-2652.
\doi{10.1086/707011}
}
\description{
European polities panel studying whether queens were more likely than
kings to engage in warfare, using gender of firstborn as an instrument.
}
\examples{
\dontrun{
queens <- grossman::load("queens")
}

}
\keyword{datasets}
