% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{thirdkid}
\alias{thirdkid}
\title{Third Kid (Family Size) Dataset}
\format{
A data frame with 254,652 rows and 34 variables:
\describe{
\item{kidcount}{Number of children}
\item{morekids}{Had more than 2 kids}
\item{boys2}{First two births boys}
\item{girls2}{First two births girls}
\item{boy1st}{First birth boy}
\item{samesex}{First two kids are of same sex}
\item{multi2nd}{Second birth is a multiple (twin)}
\item{agem}{Age of mother}
\item{agefstm}{Age of mom at first birth}
\item{blackm}{Black (mom)}
\item{hispm}{Hispanic (mom)}
\item{othracem}{Other race (mom)}
\item{workedm}{Mom worked last year}
\item{weeksm1}{Weeks worked (mom)}
\item{hourswm}{Hours of work per week (mom)}
\item{incomem}{Mom's labor income}
\item{faminc1}{Family income}
\item{nonmomi}{Income not generated by mom}
\item{famincl}{Log family income}
\item{moreths}{Mom has more than high school education}
\item{boy2nd}{Second birth boy}
\item{educm}{Mom's education (years)}
\item{hsormore}{Mom has high school or more education}
\item{hsgrad}{Mom is a high school graduate}
\item{blackd}{Black (dad)}
\item{hispd}{Hispanic (dad)}
\item{othraced}{Other race (dad)}
\item{workedd}{Dad worked last year}
\item{weeksd1}{Weeks worked (dad)}
\item{hourswd}{Hours of work per week (dad)}
\item{incomed}{Dad's labor income}
\item{ageq2nd}{Age of second kid in quarters}
\item{ageqk}{Age of first kid in quarters}
\item{ageq3rd}{Age of 3rd kid in quarters}
}
}
\source{
Angrist, Joshua D. and William N. Evans. 1998. "Children and Their
Parents' Labor Supply: Evidence from Exogenous Variation in Family Size."
\emph{American Economic Review} 88(3): 450-477.
}
\description{
1980 Census microdata on family size and labor supply for married women
with two or more children. Uses same-sex of first two children as an
instrument for having a third child.
}
\examples{
\dontrun{
thirdkid <- grossman::load("thirdkid")
}

}
\keyword{datasets}
