% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{unions}
\alias{unions}
\title{Union Wage Panel Dataset}
\format{
A data frame with 4,360 rows and 36 variables:
\describe{
\item{nr}{Person identifier}
\item{year}{Observation year}
\item{ag}{Industry: Agriculture}
\item{black}{Black}
\item{bus}{Industry: Business & Repair Service}
\item{con}{Industry: Construction}
\item{ent}{Industry: Entertainment}
\item{exper}{Experience (Age - 6 - Schooling)}
\item{fin}{Industry: Finance}
\item{hisp}{Hispanic}
\item{hlth}{Has health disability}
\item{hours}{Hours worked (year)}
\item{man}{Industry: Manufacturing}
\item{mar}{Married}
\item{min}{Industry: Mining}
\item{nc}{Lives in North Central}
\item{ne}{Lives in North East}
\item{occ1}{Occupation: Professional, Technical and kindred}
\item{occ2}{Occupation: Managers, Officials and Proprietors}
\item{occ3}{Occupation: Sales Workers}
\item{occ4}{Occupation: Clerical and kindred}
\item{occ5}{Occupation: Craftsmen, Foremen and kindred}
\item{occ6}{Occupation: Operatives and kindred}
\item{occ7}{Occupation: Laborers and farmers}
\item{occ8}{Occupation: Farm Laborers and Foreman}
\item{occ9}{Occupation: Service Workers}
\item{per}{Industry: Personal Service}
\item{pro}{Industry: Professional & Related Service}
\item{pub}{Industry: Public Administration}
\item{rur}{Lives in rural area}
\item{s}{Lives in South}
\item{school}{Years of schooling}
\item{tra}{Industry: Transportation}
\item{trad}{Industry: Trade}
\item{union}{Wage set by collective bargaining}
\item{wage}{Log of hourly wage}
}
}
\source{
Vella, Francis and Marno Verbeek. 1998. "Whose Wages Do Unions
Raise? A Dynamic Model of Unionism and Wage Rate Determination for Young
Men." \emph{Journal of Applied Econometrics} 13(2): 163-183.
\doi{10.1002/(SICI)1099-1255(199803/04)13:2<163::AID-JAE460>3.0.CO;2-Y}
}
\description{
Panel data on wages and union status for young men, used to study the
effect of unions on wages.
}
\examples{
\dontrun{
unions <- grossman::load("unions")
}

}
\keyword{datasets}
