% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{widows}
\alias{widows}
\title{Widows (Land Inheritance) Dataset}
\format{
A data frame with 7,825 rows and 16 variables:
\describe{
\item{fert}{Total fertilizer use (kg)}
\item{noninh}{Non-inheritance land access}
\item{fertsub}{Any free fertilizer received}
\item{hh_landaccess}{Household land access}
\item{numadults}{Number of adult equivalents}
\item{district}{District}
\item{hh_couple}{Household is a couple}
\item{sea_group}{SEA group}
\item{hh_number_cattle}{Number of cattle}
\item{hh_aindex}{Asset index}
\item{female_agriculture}{Female agriculture indicator}
}
}
\source{
Dillon, Brian and Alessandra Voena. 2018. "Widows' Land Rights
and Agricultural Investment." \emph{Journal of Development Economics}
135: 449-460. \doi{10.1016/j.jdeveco.2018.08.006}
}
\description{
Household-level data on land inheritance, fertilizer use, and
agricultural outcomes in Kenya.
}
\examples{
\dontrun{
widows <- grossman::load("widows")
}

}
\keyword{datasets}
