% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{bureaucrats}
\alias{bureaucrats}
\title{Bureaucrats (Village Officials) Dataset}
\format{
A data frame with 2,809 rows and 18 variables:
\describe{
\item{village}{Village ID}
\item{year}{Year}
\item{cgvo}{CGVO has been assigned}
\item{l_subsidy_rate}{Log subsidized population (per 1000)}
\item{l_poor_housing_rate}{Log poor housing households (per 100)}
\item{l_poor_reg_rate}{Log registered poor households (per 100)}
\item{l_disability_rate}{Log population with disabilities (per 1000)}
\item{district}{District ID}
\item{precipitation}{Daily average precipitation (inches)}
\item{temperature}{Daily average temperature (Fahrenheit)}
\item{Lead_D4_plus}{Relative time <= -4}
\item{Lead_D3}{Relative time -3}
\item{Lead_D2}{Relative time -2}
\item{Lead_D1}{Relative time -1}
\item{D0}{Relative time 0}
\item{Lag_D1}{Relative time 1}
\item{Lag_D2}{Relative time 2}
\item{Lag_D3_plus}{Relative time >= 3}
}
}
\source{
He, Guojun and Shaoda Wang. 2017. "Do College Graduates Serving
as Village Officials Help Rural China?" \emph{American Economic Journal:
Applied Economics} 9(4): 186-215. \doi{10.1257/app.20160079}
}
\description{
Village-year panel data from China on the assignment of College Graduate
Village Officials (CGVOs) and welfare outcomes. Used to study the effect
of improved bureaucrat quality on poverty alleviation.
}
\examples{
\dontrun{
bureaucrats <- grossman::load("bureaucrats")
}

}
\keyword{datasets}
