% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{cookstove}
\alias{cookstove}
\title{Cookstove Adoption Dataset}
\format{
A data frame with 836 rows and 6 variables:
\describe{
\item{respondent_id}{Respondent identifier}
\item{amount_weekly}{Mean weekly charcoal spending (USD)}
\item{jikokoa}{Bought cookstove (1=yes, 0=no)}
\item{wtp}{Willingness to pay (KES)}
\item{amount_weekly_pre}{Pre-adoption charcoal spending (KSH/week)}
\item{price}{Offered price (USD)}
}
}
\source{
Berkouwer, Susanna B., and Joshua T. Dean. 2022. "Credit, Attention,
and Externalities in the Adoption of Energy Efficient Technologies by
Low-Income Households." \emph{American Economic Review} 112(10): 3291-3330.
\doi{10.1257/aer.20210766}
}
\description{
Data from a cookstove adoption experiment in Nairobi, collapsed to
respondent-level from repeated SMS surveys.
}
\examples{
\dontrun{
cs <- grossman::load("cookstove")
}

}
\keyword{datasets}
