% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{hurricane}
\alias{hurricane}
\title{Hurricane Dataset}
\format{
A data frame with 49,698 rows and 7 variables:
\describe{
\item{county}{County FIPS code}
\item{year}{Year}
\item{hyear}{First hurricane year (Inf if none)}
\item{govtrans}{Log per capita government transfers}
\item{bustrans}{Log per capita business transfers}
\item{emprate}{Total adult employment rate}
\item{avgwage}{Log average wage and salary disbursements}
}
}
\source{
Deryugina, Tatyana. 2017. "The Fiscal Cost of Hurricanes: Disaster
Aid versus Social Insurance." \emph{American Economic Journal: Economic
Policy} 9(3): 168-198. \doi{10.1257/pol.20140296}
}
\description{
County-year panel on hurricane exposure and economic outcomes,
used to study the fiscal cost of hurricanes.
}
\examples{
\dontrun{
hurricane <- grossman::load("hurricane")
}

}
\keyword{datasets}
