% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{kenyagrid}
\alias{kenyagrid}
\title{Kenya Grid Electrification Dataset}
\format{
A data frame with 4,368 rows and 52 variables (alphabetically
ordered). Key variables include:
\describe{
\item{connected}{Grid connected}
\item{treated_all}{Any treatment}
\item{treat_low}{Treatment arm #1 (25k subsidy)}
\item{treat_med}{Treatment arm #2 (15k subsidy)}
\item{treat_full}{Treatment arm #3 (0\% subsidy)}
\item{emi}{Economic Index}
\item{nemi}{Non-Economic Index}
\item{hours_worked}{Total hours worked last week}
\item{hhid}{Household ID}
\item{siteno}{Community ID}
}
}
\source{
Lee, Kenneth, Edward Miguel, and Catherine Wolfram. 2020.
"Experimental Evidence on the Economics of Rural Electrification."
\emph{Journal of Political Economy} 128(4): 1523-1565.
\doi{10.1086/705417}
}
\description{
Household panel data from a rural electrification experiment in Kenya.
Randomized subsidies for grid connection to study the economics of
rural electrification.
}
\examples{
\dontrun{
kenyagrid <- grossman::load("kenyagrid")
}

}
\keyword{datasets}
