% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{kindy}
\alias{kindy}
\title{Kindergarten and Maternal Labor Supply Dataset}
\format{
A data frame with 17,817 rows and 23 variables:
\describe{
\item{youngest}{Age of youngest child}
\item{public}{Child enrolled in public school}
\item{quarter}{Quarter of birth}
\item{enrold}{Child enrolled in any school}
\item{age}{Mother's age}
\item{grade}{Mother's education}
\item{centcity}{Resides in central city}
\item{white}{Mother's race is white}
\item{num05}{Number of kids aged 0-5 in HH}
\item{num612}{Number of kids aged 6-12 in HH}
\item{num1317}{Number of kids aged 13-17 in HH}
\item{numge18}{Number of kids aged 18+ in HH}
\item{othrlt18}{Number of other HH members aged <18}
\item{othrge18}{Number of other HH members aged 18+}
\item{weeksw79}{Weeks worked in 1979}
\item{work79}{Employment in 1979}
\item{hours79}{Hours worked in 1979}
\item{worked}{Worked last week}
\item{hours}{Hours worked last week}
\item{assist}{Received public assistance in 1979}
\item{salary}{Wage and salary income in 1979}
\item{state}{State of current residence}
\item{bmodk5}{Birth state of child}
}
}
\source{
Gelbach, Jonah B. 2002. "Public Schooling for Young Children and
Maternal Labor Supply." \emph{American Economic Review} 92(1): 307-322.
\doi{10.1257/000282802760015748}
}
\description{
CPS data on kindergarten enrollment and maternal labor supply, used
to study the effect of public schooling for young children on mothers'
employment.
}
\examples{
\dontrun{
kindy <- grossman::load("kindy")
}

}
\keyword{datasets}
