% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{olyset}
\alias{olyset}
\title{Olyset Bednet Experiment Dataset}
\format{
A data frame with 1,078 rows and 4 variables:
\describe{
\item{usenet}{Uses bed net (1=yes, 0=no)}
\item{ownnet}{Owns bed net (1=yes, 0=no)}
\item{price}{Price of bed net (Kenyan shillings)}
\item{village}{Village identifier}
}
}
\source{
Dupas, Pascaline. 2014. "Short-Run Subsidies and Long-Run Adoption
of New Health Products: Evidence from a Field Experiment."
\emph{Econometrica} 82(1): 197-228. \doi{10.3982/ECTA9508}
}
\description{
Village-level data on bed net ownership and usage from a price
randomization experiment in Kenya.
}
\examples{
\dontrun{
olyset <- grossman::load("olyset")
}

}
\keyword{datasets}
