% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{psid}
\alias{psid}
\title{PSID Consumption and Earnings Panel}
\format{
A data frame with 4,566 rows and 12 variables:
\describe{
\item{id}{Household head person id}
\item{year}{Year}
\item{earn}{Earnings (household head)}
\item{cons}{Consumption (total)}
\item{cons_nd}{Consumption (non-durables)}
\item{cons_srv}{Consumption (services, non-housing)}
\item{educ}{Education (household head)}
\item{age}{Age (household head)}
\item{cons_hs}{Consumption (housing)}
\item{kids}{Number of kids in household}
\item{fsize}{Family size}
\item{earn2nd}{Earnings (2nd member)}
}

An unbalanced variant with 10,370 observations is available via
\code{grossman::load("psid", variant = "unbalanced")}.
}
\source{
Blundell, Richard, Luigi Pistaferri, and Ian Preston. 2008.
"Consumption Inequality and Partial Insurance."
\emph{American Economic Review} 98(5): 1887-1921.
\doi{10.1257/aer.98.5.1887}
}
\description{
Balanced panel from the Panel Study of Income Dynamics on household
earnings and consumption.
}
\examples{
\dontrun{
psid <- grossman::load("psid")
psid_ub <- grossman::load("psid", variant = "unbalanced")
}

}
\keyword{datasets}
