% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{queens}
\alias{queens}
\title{Queens Dataset}
\format{
A data frame with 3,586 rows and 17 variables:
\describe{
\item{kingdom}{Kingdom}
\item{year}{Year}
\item{inwar}{Any participation in interstate conflict}
\item{queen}{Queen regnant indicator}
\item{fbmale}{Firstborn is male}
\item{sister}{Number of female siblings}
\item{century}{Century indicator}
\item{decade}{Decade indicator}
\item{ruler1_name}{Name of ruler 1}
\item{ruler2_name}{Name of ruler 2 (co-ruler)}
\item{firstborn_missinggender}{Firstborn missing gender}
\item{anylegit}{Any legitimate non-missing birth-year heir}
\item{anylegit_missingbirthyear}{Any legitimate missing birth-year heir}
\item{previous_corulers_unrel}{Previous co-rulers unrelated}
\item{totsib}{Total siblings}
\item{reign_iv}{Reign cluster ID (instrument)}
\item{sibling_missinggender}{Sibling with missing gender}
}
}
\source{
Dube, Oeindrila and S.P. Harish. 2020. "Queens."
\emph{Journal of Political Economy} 128(7): 2579-2652.
\doi{10.1086/707011}
}
\description{
Panel data on European monarchs and wars, used to study whether queens
were more likely than kings to engage in interstate conflict.
}
\examples{
\dontrun{
queens <- grossman::load("queens")
}

}
\keyword{datasets}
