% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{redistribution}
\alias{redistribution}
\title{Redistribution and Mobility Perceptions Dataset}
\format{
A data frame with 9,792 rows and 105 variables:
\describe{
\item{respondent_ID}{Respondent ID}
\item{country}{Country}
\item{Treated}{Treatment indicator}
\item{q1_to_q1}{Perceived chance from Q1 to Q1}
\item{government_intervention}{Optimal level of government intervention (1-7)}
\item{income_tax_top1}{Preferred tax rate, top 1\\%}
\item{estate_tax_support}{Support for estate tax}
}
}
\source{
Alesina, Alberto, Stefanie Stantcheva, and Edoardo Teso. 2018.
"Intergenerational Mobility and Preferences for Redistribution."
\emph{American Economic Review} 108(2): 521-554.
\doi{10.1257/aer.20162015}
}
\description{
Survey experiment studying how perceptions of intergenerational mobility
affect preferences for redistribution, across five countries.
}
\examples{
\dontrun{
redist <- grossman::load("redistribution")
}

}
\keyword{datasets}
