% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{thirdkid}
\alias{thirdkid}
\title{Third Kid (Same-Sex IV) Dataset}
\format{
A data frame with 254,652 rows and 34 variables:
\describe{
\item{kidcount}{Number of children}
\item{morekids}{Had more than 2 kids}
\item{samesex}{First two kids are of same sex (instrument)}
\item{workedm}{Mom worked last year}
\item{incomem}{Mother's labor income}
\item{weeksm1}{Weeks worked last year (mom)}
\item{hourswm}{Usual hours worked per week (mom)}
\item{faminc1}{Family income}
}
}
\source{
Angrist, Joshua D., and William N. Evans. 1998. "Children and
Their Parents' Labor Supply: Evidence from Exogenous Variation in Family
Size." \emph{American Economic Review} 88(3): 450-477.
}
\description{
Census data on family size and labor supply, using same-sex composition
of first two children as an instrument for having a third child.
}
\examples{
\dontrun{
tk <- grossman::load("thirdkid")
}

}
\keyword{datasets}
