% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{unions}
\alias{unions}
\title{Union Wage Premium Panel (NLSY)}
\format{
A data frame with 4,360 rows and 36 variables:
\describe{
\item{nr}{Person identifier}
\item{year}{Observation year}
\item{union}{Wage set by collective bargaining}
\item{wage}{Log hourly wage}
\item{hours}{Hours worked (year)}
\item{exper}{Experience (Age - 6 - Schooling)}
\item{school}{Years of schooling}
\item{black}{Black}
\item{hisp}{Hispanic}
\item{mar}{Married}
}
}
\source{
Vella, Francis, and Marno Verbeek. 1998. "Whose Wages Do Unions
Raise? A Dynamic Model of Unionism and Wage Rate Determination for Young
Men." \emph{Journal of Applied Econometrics} 13(2): 163-183.
}
\description{
Panel data from the National Longitudinal Survey of Youth on union
membership, wages, and individual characteristics.
}
\examples{
\dontrun{
unions <- grossman::load("unions")
}

}
\keyword{datasets}
