% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{widows}
\alias{widows}
\title{Widows, Inheritance, and Fertility Dataset}
\format{
A data frame with 7,825 rows and 16 variables:
\describe{
\item{fert}{Total fertilizer applied (kgs)}
\item{noninh}{No inheritance (=1)}
\item{fertsub}{Fertilizer subsidy available (=1)}
\item{hh_landaccess}{Total land area (Ha)}
\item{numadults}{Adult equivalents}
\item{district}{District}
\item{hh_couple}{Couples (=1)}
}
}
\source{
Voena, Alessandra. 2015. "Yours, Mine, and Ours: Do Divorce Laws
Affect the Intertemporal Behavior of Married Couples?"
\emph{American Economic Review} 105(8): 2295-2332.
\doi{10.1257/aer.20120234}
}
\description{
Zambian household data on inheritance customs, fertility, and agricultural
investment decisions.
}
\examples{
\dontrun{
widows <- grossman::load("widows")
}

}
\keyword{datasets}
